PyQT4 folder added
While searching for a way to build on Arunmozhi's code base, I found the
following project on Launchpad by Benjamin Thyreau - An application to
easily read Wikipedia's downloaded dump files:
https://launchpad.net/wikipediadumpreader

As you can see from the above link, this is open source and dual
licensed under Simplified BSD Licence and GNU GPL v2.

By combining features from Benjamin's code and Arunmozhi's code, I have
built an application that seems to do the basic functions OK. Please see
the enclosed screenshot.

Benjamin's code base had two things that we are looking for: PyQT4 user
interface and more usable (though not complete) parsing of the wiki
markup. On top of that he had built the ability to follow links. A user
can click on hyperlinked words in the results to look-up those words
further. However, on additional  testing I discovered his code base had
one big limitation for our use. It can only be used as an English to
Tamil dictionary, but Tamil words cannot be looked up. This is because
his indexing was not in unicode - as he himself noted in comments in his
code.

Arunmozhi uses Python Whoosh module for indexing and searching. Whoosh
is natively built to handle Unicode. He also split the larger Wiktionary
dump file into smaller chunks for faster look-ups. And he went a step
further and built a Windows exe as well.

"Wouldn't it be great if we can combine Benjamin's PyQT4 user interface
and wiki parsing with Arunmozhi's indexing/searching and then build a
Windows exe following Arunmozhi's steps", I thought. However, it turned
out to be harder than it initially appeared ( Isn't it always ;-) ).
Especially because I am new to Python! But, long story short, I have the
modified code as well as the Windows exe now.

The following are the only files needed for the exe to run:
1) Karthika.exe
2) Index folder "indexdir"
3) Wiktionary dump broken up in the "chunks" folder.